refactor(http): move h1 and h2 into http module
