+++ src/lib.rs
@@ -186,10 +186,6 @@ pub mod server;
 pub mod status;
 pub mod uri;
 pub mod version;
-pub mod message;
-pub mod http11;
-pub mod http2;
-
 
 /// Re-exporting the mime crate, for convenience.
 pub mod mime {
